+++ templates/repo/home.tmpl
@@ -68,7 +68,13 @@
 				{{$l := Eval $n "-" 1}}
 				<!-- If home page, show new pr. If not, show breadcrumb -->
 				{{if and (eq $n 0) .CanCompareOrPull .IsViewBranch (not .Repository.IsArchived)}}
-					<a id="new-pull-request" role="button" class="ui compact basic button" href="{{CompareLink .BaseRepo .Repository .BranchName}}"
+					{{$cmpBranch := ""}}
+					{{if ne .Repository.ID .BaseRepo.ID}}
+						{{$cmpBranch = printf "%s/%s:" (.Repository.OwnerName|PathEscape) (.Repository.Name|PathEscape)}}
+					{{end}}
+					{{$cmpBranch = printf "%s%s" $cmpBranch (.BranchName|PathEscapeSegments)}}
+					{{$compareLink := printf "%s/compare/%s...%s" .BaseRepo.Link (.BaseRepo.DefaultBranch|PathEscapeSegments) $cmpBranch}}
+					<a id="new-pull-request" role="button" class="ui compact basic button" href="{{$compareLink}}"
 						data-tooltip-content="{{if .PullRequestCtx.Allowed}}{{.locale.Tr "repo.pulls.compare_changes"}}{{else}}{{.locale.Tr "action.compare_branch"}}{{end}}">
 						{{svg "octicon-git-pull-request"}}
 					</a>
@@ -105,14 +111,14 @@
 				{{if ne $n 0}}
 					<span class="ui breadcrumb repo-path gt-ml-2">
 						<a class="section" href="{{.RepoLink}}/src/{{.BranchNameSubURL}}" title="{{.Repository.Name}}">{{StringUtils.EllipsisString .Repository.Name 30}}</a>
-						{{range $i, $v := .TreeNames}}
+						{{- range $i, $v := .TreeNames -}}
 							<span class="divider">/</span>
-							{{if eq $i $l}}
+							{{- if eq $i $l -}}
 								<span class="active section" title="{{$v}}">{{StringUtils.EllipsisString $v 30}}</span>
-							{{else}}
+							{{- else -}}
 								{{$p := index $.Paths $i}}<span class="section"><a href="{{$.BranchLink}}/{{PathEscapeSegments $p}}" title="{{$v}}">{{StringUtils.EllipsisString $v 30}}</a></span>
-							{{end}}
-						{{end}}
+							{{- end -}}
+						{{- end -}}
 					</span>
 				{{end}}
 			</div>
